moodify-backend/requirements.txt
flask
spotipy
python-dotenv
flask-cors
spacy
transformers
torch
requests
